requirements.txt
azure-ai-formrecognizer==3.3.3
azure-ai-ml==1.22.4
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
azure-storage-file-share==12.20.0
chainlit==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
openai==1.55.0
psutil==6.1.0
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
streamlit==1.40.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.25.0
